Added contextual policies
