+++ genrl/classical/bandit/contextual_policies/__init__.py
@@ -1,0 +1,13 @@
+from genrl.classical.bandit.contextual_policies.bayesian import (  # noqa
+    BayesianUCBCBPolicy,
+)
+from genrl.classical.bandit.contextual_policies.epsgreedy import (  # noqa
+    EpsGreedyCBPolicy,
+)
+from genrl.classical.bandit.contextual_policies.gradient import GradientCBPolicy  # noqa
+from genrl.classical.bandit.contextual_policies.thompson import (  # noqa
+    ThompsonSamplingCBPolicy,
+)
+from genrl.classical.bandit.contextual_policies.ucb import UCBCBPolicy  # noqa
+
+from genrl.classical.bandit.contextual_policies.base import CBPolicy  # noqa; noqa; noqa
