+++ genrl/classical/bandit/contextual_policies/bayesian.py
@@ -1,11 +1,11 @@
 import numpy as np
 from scipy import stats
 
-from ...bandit import Bandit
-from .base import BanditPolicy
+from ..contextual_bandits import ContextualBandit
+from .base import CBPolicy
 
 
-class BayesianUCBPolicy(BanditPolicy):
+class BayesianUCBCBPolicy(CBPolicy):
     """
     Multi-Armed Bandit Solver with Bayesian Upper Confidence Bound
     based Action Selection Strategy.
@@ -15,24 +15,24 @@ class BayesianUCBPolicy(BanditPolicy):
     :param bandit: The Bandit to solve
     :param alpha: alpha value for beta distribution
     :param beta: beta values for beta distibution
-    :param confidence: Confidence level which controls degree of exploration
-    :type bandit: Bandit type object
+    :param c: Confidence level which controls degree of exploration
+    :type bandit: ContextualBandit type object
     :type alpha: float
     :type beta: float
-    :type confidence: float
+    :type c: float
     """
 
     def __init__(
         self,
-        bandit: Bandit,
+        bandit: ContextualBandit,
         alpha: float = 1.0,
         beta: float = 1.0,
         confidence: float = 3.0,
     ):
-        super(BayesianUCBPolicy, self).__init__(bandit)
+        super(BayesianUCBCBPolicy, self).__init__(bandit)
         self._c = confidence
-        self._a = alpha * np.ones(self._bandit.arms)
-        self._b = beta * np.ones(self._bandit.arms)
+        self._a = alpha * np.ones(shape=(bandit.bandits, bandit.arms))
+        self._b = beta * np.ones(shape=(bandit.bandits, bandit.arms))
 
     @property
     def quality(self) -> np.ndarray:
@@ -74,41 +74,46 @@ class BayesianUCBPolicy(BanditPolicy):
         """
         return self._c
 
-    def select_action(self, timestep: int) -> int:
+    def select_action(self, context: int, t: int) -> int:
         """
         Select an action according to bayesian upper confidence bound
 
-        Take action that maximises a weighted sum of the quality values and
+        Take action that maximises a weighted sum of the Q values and
         a beta distribution paramerterized by alpha and beta
-        and weighted by c for each action.
+        and weighted by c for each action
 
-        :param timestep: timestep to choose action for
-        :type timestep: int
+        :param context: the context to select action for
+        :param t: timestep to choose action for
+        :type context: int
+        :type t: int
         :returns: Selected action
         :rtype: int
         """
         action = np.argmax(
-            self.quality + stats.beta.std(self.a, self.b) * self.confidence
+            self.quality[context]
+            + stats.beta.std(self.a[context], self.b[context]) * self.confidence
         )
-        self.action_hist.append(action)
+        self.action_hist.append((context, action))
         return action
 
-    def update_params(self, action: int, reward: float) -> None:
+    def update_params(self, context: int, action: int, reward: float) -> None:
         """
         Update parmeters for the policy
 
-        Updates the regret as the difference between max quality value and
-        that of the action. Updates the quality values according to the
-        reward recieved in this step.
+        Updates the regret as the difference between max Q value and
+        that of the action. Updates the Q values according to the
+        reward recieved in this step
 
+        :param context: context for which action is taken
         :param action: action taken for the step
         :param reward: reward obtained for the step
+        :type context: int
         :type action: int
         :type reward: float
         """
         self.reward_hist.append(reward)
-        self.a[action] += reward
-        self.b[action] += 1 - reward
-        self._regret += max(self.quality) - self.quality[action]
+        self.a[context, action] += reward
+        self.b[context, action] += 1 - reward
+        self._regret += max(self.quality[context]) - self.quality[context, action]
         self.regret_hist.append(self.regret)
-        self.counts[action] += 1
+        self.counts[context, action] += 1
