+++ genrl/classical/bandit/policies/__init__.py
@@ -1,6 +1,6 @@
 from genrl.classical.bandit.policies.bayesian import BayesianUCBPolicy  # noqa
 from genrl.classical.bandit.policies.epsgreedy import EpsGreedyPolicy  # noqa
-from genrl.classical.bandit.policies.softmax import SoftmaxActionSelectionPolicy  # noqa
+from genrl.classical.bandit.policies.gradient import GradientPolicy  # noqa
 from genrl.classical.bandit.policies.thompson import ThompsonSamplingPolicy  # noqa
 from genrl.classical.bandit.policies.ucb import UCBPolicy  # noqa
