+++ requirements.txt
@@ -11,8 +11,8 @@ scipy==1.4.1
 six==1.14.0
 matplotlib==3.2.1
 pytest==5.4.1
-torch==1.4.0
-torchvision==0.5.0
+#torch==1.4.0
+#torchvision==0.5.0
 tensorflow-tensorboard==1.5.1
 tensorboard==1.15.0
 pre-commit==2.4.0
